move cyno's a1 dmg buff to asc.go
Also changed/removed redundant comments.
